requirements.txt
requests==2.31.0
urllib3==2.0.3
colorama==0.4.6
discord.py==2.3.1
Pillow==10.0.0
python-dotenv
aiohttp